feat(lint): add doc_params_mismatch lint
