+++ Cargo.toml
@@ -11,6 +11,7 @@ keywords = ["ronat", "lint", "plugin", "documentation", "spelling"]
 [dependencies]
 ispell = "0.3.0"
 markdown = "0.2.0"
+docstrings = {git = "https://github.com/killercup/rust-docstrings.git"}
 
 [dev-dependencies]
 compiletest_rs = "0.2.5"
